Separate visual and app layer in customers
