+++ src/customers/components/CustomerListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./CustomerListPage";
 export * from "./CustomerListPage";
+export * from "./filters";
