+++ src/customers/views/CustomerList/CustomerList.tsx
@@ -44,7 +44,7 @@ import {
   saveFilterTab,
   getFilterQueryParam,
   getFilterOpts
-} from "./filter";
+} from "./filters";
 import { getSortQueryVariables } from "./sort";
 
 interface CustomerListProps {
