--- src/customers/views/CustomerList/messages.ts
@@ -1,17 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  joinDate: {
-    defaultMessage: "Join Date",
-    description: "customer"
-  },
-  moneySpent: {
-    defaultMessage: "Money Spent",
-    description: "customer"
-  },
-  numberOfOrders: {
-    defaultMessage: "Number of Orders"
-  }
-});
-
-export default messages;
